Create `Debounce` object by duration
